+++ src/main/java/com/alfy/budget/model/BudgetInfo.java
@@ -1,10 +1,11 @@
 package com.alfy.budget.model;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class BudgetInfo {
 
-    public double total;
+    public BigDecimal total;
     public int percent;
     public Budget budget;
     public List<CategoryInfo> categories;
